split out merkle_proof.rs from pmmr.rs split out tests in core as well
