+++ chain/src/txhashset.rs
@@ -25,9 +25,9 @@ use std::time::Instant;
 use util::secp::pedersen::{Commitment, RangeProof};
 
 use core::consensus::REWARD;
-use core::core::{Block, BlockHeader, Committed, Input, Output, OutputFeatures, OutputIdentifier,
+use core::core::{Block, BlockHeader, Committed, Input, MerkleProof, Output, OutputFeatures, OutputIdentifier,
                  TxKernel};
-use core::core::pmmr::{self, MerkleProof, PMMR};
+use core::core::pmmr::{self, PMMR};
 use core::global;
 use core::core::hash::{Hash, Hashed};
 use core::ser::{PMMRIndexHashable, PMMRable};
