+++ wallet/src/types.rs
@@ -35,9 +35,8 @@ use tokio_retry::strategy::FibonacciBackoff;
 use failure::{Backtrace, Context, Fail, ResultExt};
 
 use core::consensus;
-use core::core::Transaction;
+use core::core::{MerkleProof, Transaction};
 use core::core::hash::Hash;
-use core::core::pmmr::MerkleProof;
 use core::ser;
 use keychain;
 use util;
